refactor: Change processor name
