+++ src/ai/backend/manager/models/endpoint.py
@@ -77,7 +77,7 @@ from .vfolder import VFolderRow, prepare_vfolder_mounts
 
 if TYPE_CHECKING:
     from ai.backend.manager.config import SharedConfig
-    from ai.backend.manager.services.model_service.types import EndpointTokenData
+    from ai.backend.manager.services.model_serving.types import EndpointTokenData
 
     from .gql import GraphQueryContext
 
@@ -632,7 +632,7 @@ class EndpointTokenRow(Base):
         self.session_owner = user_uuid
 
     def to_dataclass(self) -> EndpointTokenData:
-        from ai.backend.manager.services.model_service.types import EndpointTokenData
+        from ai.backend.manager.services.model_serving.types import EndpointTokenData
 
         return EndpointTokenData(
             id=self.id,
