--- src/ai/backend/manager/services/model_service/actions/create_auto_scaling_rule.py
@@ -1,36 +1,0 @@
-from dataclasses import dataclass
-from typing import Optional, override
-
-from ai.backend.common.types import EndpointId
-from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.model_service.actions.base import ModelServiceAction
-from ai.backend.manager.services.model_service.types import (
-    EndpointAutoScalingRuleCreator,
-    EndpointAutoScalingRuleData,
-    RequesterCtx,
-)
-
-
-@dataclass
-class CreateEndpointAutoScalingRuleAction(ModelServiceAction):
-    requester_ctx: RequesterCtx
-    endpoint_id: EndpointId
-    creator: EndpointAutoScalingRuleCreator
-
-    @override
-    def entity_id(self) -> Optional[str]:
-        return None
-
-    @override
-    def operation_type(self) -> str:
-        return "create"
-
-
-@dataclass
-class CreateEndpointAutoScalingRuleActionResult(BaseActionResult):
-    success: bool
-    data: Optional[EndpointAutoScalingRuleData]
-
-    @override
-    def entity_id(self) -> Optional[str]:
-        return str(self.data.id) if self.data is not None else None
