+++ src/ai/backend/manager/services/model_serving/actions/force_sync.py
@@ -3,27 +3,24 @@ from dataclasses import dataclass
 from typing import Optional, override
 
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.model_service.actions.base import ModelServiceAction
-from ai.backend.manager.services.model_service.types import RequesterCtx
+from ai.backend.manager.services.model_serving.actions.base import ModelServiceAction
+from ai.backend.manager.services.model_serving.types import RequesterCtx
 
 
 @dataclass
-class DeleteRouteAction(ModelServiceAction):
-    requester_ctx: RequesterCtx
+class ForceSyncAction(ModelServiceAction):
     service_id: uuid.UUID
-    route_id: uuid.UUID
+    requester_ctx: RequesterCtx
 
-    @override
     def entity_id(self) -> Optional[str]:
         return None
 
-    @override
     def operation_type(self) -> str:
-        return "delete"
+        return "sync"
 
 
 @dataclass
-class DeleteRouteActionResult(BaseActionResult):
+class ForceSyncActionResult(BaseActionResult):
     success: bool
 
     @override
