+++ src/ai/backend/manager/services/model_serving/actions/generate_token.py
@@ -6,8 +6,8 @@ from typing import Optional, override
 from dateutil.relativedelta import relativedelta
 
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.model_service.actions.base import ModelServiceAction
-from ai.backend.manager.services.model_service.types import EndpointTokenData, RequesterCtx
+from ai.backend.manager.services.model_serving.actions.base import ModelServiceAction
+from ai.backend.manager.services.model_serving.types import EndpointTokenData, RequesterCtx
 
 
 @dataclass
