+++ src/ai/backend/manager/services/model_serving/actions/list_model_service.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 from typing import Optional, override
 
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.model_service.actions.base import ModelServiceAction
-from ai.backend.manager.services.model_service.types import CompactServiceInfo
+from ai.backend.manager.services.model_serving.actions.base import ModelServiceAction
+from ai.backend.manager.services.model_serving.types import CompactServiceInfo
 
 
 @dataclass
