+++ src/ai/backend/manager/services/model_serving/actions/update_route.py
@@ -3,14 +3,16 @@ from dataclasses import dataclass
 from typing import Optional, override
 
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.model_service.actions.base import ModelServiceAction
-from ai.backend.manager.services.model_service.types import RequesterCtx
+from ai.backend.manager.services.model_serving.actions.base import ModelServiceAction
+from ai.backend.manager.services.model_serving.types import RequesterCtx
 
 
 @dataclass
-class DeleteModelServiceAction(ModelServiceAction):
-    service_id: uuid.UUID
+class UpdateRouteAction(ModelServiceAction):
     requester_ctx: RequesterCtx
+    service_id: uuid.UUID
+    route_id: uuid.UUID
+    traffic_ratio: float
 
     @override
     def entity_id(self) -> Optional[str]:
@@ -18,11 +20,11 @@ class DeleteModelServiceAction(ModelServiceAction):
 
     @override
     def operation_type(self) -> str:
-        return "delete"
+        return "update"
 
 
 @dataclass
-class DeleteModelServiceActionResult(BaseActionResult):
+class UpdateRouteActionResult(BaseActionResult):
     success: bool
 
     @override
