+++ src/ai/backend/manager/services/model_serving/services/BUILD
