+++ src/ai/backend/manager/services/model_serving/services/utils.py
@@ -1,11 +1,11 @@
 import uuid
 
 from ai.backend.manager.models.utils import ExtendedAsyncSAEngine
-from ai.backend.manager.services.model_service.exceptions import (
+from ai.backend.manager.services.model_serving.exceptions import (
     GenericForbidden,
     InvalidAPIParameters,
 )
-from ai.backend.manager.services.model_service.types import RequesterCtx
+from ai.backend.manager.services.model_serving.types import RequesterCtx
 from ai.backend.manager.utils import check_if_requester_is_eligible_to_act_as_target_user_uuid
