+++ src/ai/backend/manager/services/model_serving/types.py
